--- xBox/src/cmd/AdminCheckPay.java
@@ -1,5 +1,0 @@
-package cmd;
-
-public class AdminCheckPay {
-    
-}
