+++ xBox/src/cmd/AdminPayApprove.java
@@ -16,19 +16,23 @@ public class AdminPayApprove extends Undoable{
 
     public void execute(String[] cmdLine){
         /*
-         * adminName of the reuqest,clientName
+         * adminName of the reuqest,clientName，payOrNot
         */
-        if(cmdLine[0]=="ubox@gmail.com"){
-            RequestSearcher requestSearcher=RequestSearcher.getInstance();
-            Client thisClient=clientSearcher.searchByClientEmail(cmdLine[0]);
-            ArrayList<Request> request= requestSearcher.searchByClient(thisClient);
-            for (Request request2:request){
-                request2.getTarget().changeRentableStatus(request2.getRentable());;
-                requestManager.removeRequest(request2);
+        RequestSearcher requestSearcher=RequestSearcher.getInstance();
+
+        if(cmdLine[0].equals("ubox@gmail.com")){
+             if(cmdLine[2].equals("paied")){
+                
+                Client thisClient=clientSearcher.searchByClientEmail(cmdLine[0]);
+                ArrayList<Request> request= requestSearcher.searchByClient(thisClient);
+                for (Request request2:request){
+                    request2.getTarget().changeRentableStatus(request2.getRentable());;
+                    requestManager.removeRequest(request2);
+                }
+                recordManager.insert(thisClient,rentable,SystemDate.toDate(cmdLine[2])，);
+            }else if(cmdLine[2].equals("unpaied")){
+
             }
-        }
-        else{
-            
         }
     }
 }
