--- xBox/src/cmd/PayBoxCommand.java
@@ -1,16 +1,0 @@
-package cmd;
-
-public class PayBoxCommand extends Undoable{
-    @Override
-    public void undo(){
-
-    }
-    @Override
-    public void redo(){
-
-    }
-
-    public void execute(String[] cmdLine){
-        
-    }
-}
